beginnings of the process for calculating leisure FOI in
briefly samples whether an individual goes to a leisure setting or stays home on a given timestep. Then, if out, samples which leisure setting. Then uses this to create a new CategoricalVariable for that day's leisure settings visited, which we use to calculate FOI.

NOTE it's currently doing this every timestep rather than daily. We should change that behaviour somehow.