Split population generation from running the model.
Generating the population data (eg. assigning individuals to locations)
can take significant amounts of time. By refactoring the interface a
little bit, we can split off the generation step into its own function,
which allows it to be re-used, saving time along the way.

This is especially useful during development or benchmarking of the main
model. For producing scientifically useful results it may be preferable
to keep generating population data each time.

By default `run_simulation` will generate a new population each time it
is called, but it can be passed a `population_data` argument to avoid
this.  The `population_data` is obtained by calling
`generate_population_data`.

The population data object is free of any objects from the individual
package: it has normal value semantics and it can be serialized to a
file.

I've imported strict_list from [reside.utils](reside.utils) and used it
to define the parameter and variable list. It made refactoring much
easier and I think is a good addition, though it could of course be
removed.

[reside.utils]: https://github.com/reside-ic/reside.utils